Enable source map support, improve README
