+++ packages/api/package.json
@@ -7,7 +7,7 @@
   },
   "main": "index.js",
   "scripts": {
-    "dev": "pnpm ts-node src/local-server.ts",
+    "dev": "pnpm ts-node src/index.ts",
     "docker:build": "docker compose --file docker/docker-compose.yml build",
     "docker:start": "docker compose --file docker/docker-compose.yml up",
     "docker:start:detach": "docker compose --file docker/docker-compose.yml up --detach",
@@ -22,7 +22,8 @@
   "license": "MIT",
   "devDependencies": {
     "@types/express": "^4.17.17",
-    "@types/lodash": "^4.14.194"
+    "@types/lodash": "^4.14.194",
+    "@types/source-map-support": "^0.5.7"
   },
   "dependencies": {
     "@api3/promise-utils": "0.3.0",
@@ -30,6 +31,7 @@
     "ethers": "^5.7.2",
     "express": "^4.18.2",
     "lodash": "^4.17.21",
+    "source-map-support": "^0.5.21",
     "zod": "^3.21.4"
   }
 }
