+++ packages/data-pusher/package.json
@@ -37,6 +37,7 @@
   },
   "devDependencies": {
     "@types/express": "^4.17.13",
-    "@types/lodash": "^4.14.186"
+    "@types/lodash": "^4.14.186",
+    "@types/source-map-support": "^0.5.7"
   }
 }
